Refactor SFTP transfer models
