+++ lib/view/page/storage/sftp_mission.dart
@@ -5,7 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:server_box/core/extension/context/locale.dart';
 import 'package:server_box/core/utils/refresh_interval.dart';
-import 'package:server_box/data/model/sftp/worker.dart';
+import 'package:server_box/data/model/sftp/req.dart';
+import 'package:server_box/data/model/sftp/status.dart';
 import 'package:server_box/data/provider/sftp.dart';
 import 'package:server_box/data/res/default.dart';
 import 'package:server_box/view/page/storage/local.dart';
